#2 TNSController, ViewController, MessageHandler 추가
