--- TNS_Server/TNS.cpp
@@ -1,11 +1,0 @@
-#include "TNS.h"
-
-int main() {
-	// 서버 인스턴트
-	TCPSocket server;
-
-	// 서버시작
-	server.StartServer();
-
-	return 0;
-}
