+++ TNS_Server/TNS_Server.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TNS_Server", "TNS_Server\TNS_Server.vcxproj", "{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Debug|x64.ActiveCfg = Debug|x64
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Debug|x64.Build.0 = Debug|x64
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Debug|x86.ActiveCfg = Debug|Win32
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Debug|x86.Build.0 = Debug|Win32
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Release|x64.ActiveCfg = Release|x64
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Release|x64.Build.0 = Release|x64
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Release|x86.ActiveCfg = Release|Win32
+		{96B3FF39-6AAC-4BF2-A96D-BF6BF6F42022}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
