+++ TNS_Server/TNS_Server/Debug/TNS_Server.tlog/TNS_Server.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
+Debug|Win32|C:\Users\JunYoung\Documents\카카오톡 받은 파일\Projects_1008\TNS_Server\|
