+++ TNS_Server/TNS_Server/TNS.h
@@ -1,0 +1,14 @@
+#include <stdlib.h>
+#include <iostream>
+#include <stdio.h>
+#include <string.h>
+#include <winsock2.h> 
+
+#pragma once
+
+using namespace std;
+#include "TopicNameTable.h"
+#include "TCPSocket.h"
+#include "RequestTable.h"
+
+TCPSocket socketServer;
