+++ TNS_Server/TNS_Server/TNSController.h
@@ -1,0 +1,32 @@
+#include "stdafx.h"
+#include"ViewManager.h"
+#include"SocketManager.h"
+#include "MessageHandler.h"
+#include <process.h>
+#include <deque>
+#include <list>
+
+#pragma once
+
+class TNSController
+{
+public:
+	TNSController();
+	~TNSController();
+
+	void	startTNSServer();
+
+private:
+	SocketManager		* socketManager;
+
+	MessageHandler	* messageHandler;
+
+	CRITICAL_SECTION	cs;
+	deque<pair<IN_ADDR, PDD_NODE>>	* recvData;
+
+	void	initalizeSetting();
+	void	closeTNSServer();
+	bool	isReceviedDataExist();
+
+	void	inputDummyDataToDB();
+};
