+++ TNS_Server/TNS_Server/ViewManager.cpp
@@ -1,0 +1,11 @@
+#include "ViewManager.h"
+
+ViewManager::ViewManager()
+{
+
+}
+
+ViewManager::~ViewManager()
+{
+
+}
