+++ TNS_Server/TNS_Server/ViewManager.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class ViewManager
+{
+	ViewManager();
+	~ViewManager();
+};
